rag-battle-3/requirements.txt
groundx-python-sdk
langchain
langchain-community
langchain-openai
openai
pinecone-client
tiktoken
llama-index
llama-index-vector-stores-pinecone
torch
transformers
python-pptx
Pillow
docx2txt
unstructured[all-docs]
poppler-utils
pytesseract
python-dotenv